requirements.txt
gate-api>=7.1.8
ollama>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pandas>=2.0.0
numpy>=1.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0